+++ include/vrv/liquescent.h
@@ -29,20 +29,20 @@ public:
     ///@{
     Liquescent();
     virtual ~Liquescent();
-    virtual Object *Clone() const { return new Liquescent(*this); }
-    virtual void Reset();
-    virtual std::string GetClassName() const { return "Liquescent"; }
+    Object *Clone() const override { return new Liquescent(*this); }
+    void Reset() override;
+    std::string GetClassName() const override { return "Liquescent"; }
     ///@}
 
     /**
      * @name Getter to interfaces
      */
     ///@{
-    virtual PitchInterface *GetPitchInterface() { return dynamic_cast<PitchInterface *>(this); }
+    PitchInterface *GetPitchInterface() override { return dynamic_cast<PitchInterface *>(this); }
     ///@}
 
     /** Override the method since alignment is required */
-    virtual bool HasToBeAligned() const { return true; }
+    bool HasToBeAligned() const override { return true; }
 
 private:
     //