+++ include/vrv/vrv.h
@@ -22,6 +22,7 @@
 #include <time.h>
 #endif
 
+#include "attdef.h"
 #include "atttypes.h"
 #include "toolkitdef.h"
 
@@ -87,6 +88,11 @@ std::string UTF32to8(const std::u32string &in);
 std::u32string UTF8to32(const std::string &in);
 
 /**
+ * Utility for converting UTF16 (std::u16string) to UTF-8
+ */
+std::string UTF16to8(const std::u16string &in);
+
+/**
  * Format a string using vsnprintf.
  * The maximum length is giving by STRING_FORMAT_MAX_LEN
  */
@@ -126,6 +132,12 @@ std::string FromCamelCase(const std::string &s);
  */
 std::string ToCamelCase(const std::string &s);
 
+/*
+ * Min / Max for data_DURATION (std::min/max not possible)
+ */
+data_DURATION DurationMin(data_DURATION dur1, data_DURATION dur2);
+data_DURATION DurationMax(data_DURATION dur1, data_DURATION dur2);
+
 /**
  *
  */
@@ -145,7 +157,7 @@ extern bool loggingToBuffer;
  */
 extern struct timeval start;
 void LogElapsedTimeStart();
-void LogElapsedTimeEnd(const char *msg = "unspecified operation");
+void LogElapsedTimeStop(const char *msg = "unspecified operation");
 
 //----------------------------------------------------------------------------
 // Notation type checks