+++ src/scoredefinterface.cpp
@@ -41,8 +41,6 @@ ScoreDefInterface::ScoreDefInterface()
     this->RegisterInterfaceAttClass(ATT_DURATIONDEFAULT);
     this->RegisterInterfaceAttClass(ATT_LYRICSTYLE);
     this->RegisterInterfaceAttClass(ATT_MEASURENUMBERS);
-    this->RegisterInterfaceAttClass(ATT_METERSIGDEFAULTLOG);
-    this->RegisterInterfaceAttClass(ATT_METERSIGDEFAULTVIS);
     this->RegisterInterfaceAttClass(ATT_MIDITEMPO);
     this->RegisterInterfaceAttClass(ATT_MMTEMPO);
     this->RegisterInterfaceAttClass(ATT_MULTINUMMEASURES);